Refactor: Complete hexagonal architecture refactor
- Renamed redis_ ports to persistence_ ports for better abstraction
- Moved memory module from context/adapters to dedicated memory/ directory
- Updated all import paths to reflect new structure
- Fixed persistence port class names and references
- All 307 tests passing with proper hexagonal separation
- Applied linting fixes and maintained code quality standards

This refactor improves the hexagonal architecture by:
- Making persistence layer technology-agnostic
- Better separation of concerns
- Cleaner import structure
- Enhanced maintainability
