+++ src/swe_ai_fleet/memory/ports/persistence_pipeline_port.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from typing import Any, Protocol
+
+
+class PersistencePipelinePort(Protocol):
+    def set(self, key: str, value: str, ex: int | None = None) -> Any: ...  # noqa: B032
+    def lpush(self, key: str, value: str) -> Any: ...
+    def expire(self, key: str, seconds: int) -> Any: ...
+    def execute(self) -> list[Any]: ...
